Add Common.vcxproj and move unicode.* and utilities.h into it.
