+++ vnext/Common/unicode.cpp
