+++ vnext/Common/unicode.h
