+++ vnext/Common/utilities.h
