+++ vnext/include/Include.vcxitems
@@ -33,7 +33,6 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)ReactWindowsCore\II18nModule.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)ReactWindowsCore\ReactWindowsAPI.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)ReactWindowsCore\SourceCodeModule.h" />
-    <ClInclude Include="$(MSBuildThisFileDirectory)ReactWindowsCore\unicode.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)ReactWindowsCore\ViewManager.h" />
   </ItemGroup>
 </Project>
